AO - refactored packages into descriptive seperation of concerns
