--- src/edu/ucsb/cs56/projects/games/pong/InstructionsListener.java
@@ -1,10 +1,0 @@
-package edu.ucsb.cs56.projects.games.pong;
-
-import java.awt.event.MouseListener;
-
-/**
- * Created by angelortega on 2/22/16.
- */
-public interface InstructionsListener extends MouseListener {
-
-}
