--- src/edu/ucsb/cs56/projects/games/pong/UserRetrievable.java
@@ -1,8 +1,0 @@
-package edu.ucsb.cs56.projects.games.pong;
-
-/**
- * Created by angel on 2/20/16.
- */
-public interface UserRetrievable {
-    String getTitle() ;
-}
