+++ src/edu/ucsb/cs56/projects/games/pong/gameplay/Game.java
@@ -1,4 +1,6 @@
-package edu.ucsb.cs56.projects.games.pong;
+package edu.ucsb.cs56.projects.games.pong.gameplay;
+
+import edu.ucsb.cs56.projects.games.pong.menu.MainMenuUI;
 
 import javax.swing.*;
 import java.awt.*;
@@ -6,11 +8,11 @@ import java.awt.*;
 /**
  * Created by angel on 2/21/16.
  */
-public class GameWindow extends JFrame {
+public class Game extends JFrame {
     private static final Dimension DIMENSION = new Dimension(640,480);
 
 
-    public GameWindow(String title, Dimension dimension) {
+    public Game(String title, Dimension dimension) {
         super(title);
         this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         this.setSize(dimension);
@@ -19,9 +21,9 @@ public class GameWindow extends JFrame {
     }
 
     public static void main(String[] args) {
-        GameWindow window = new GameWindow("Main Menu",DIMENSION);
-        MainMenuUI gameMenu = new MainMenuUI();
-        window.add(gameMenu);
+        Game window = new Game("Main Menu",DIMENSION);
+        MainMenuUI game = new MainMenuUI();
+        window.add(game);
         window.setVisible(true);
     }
 }
