+++ src/edu/ucsb/cs56/projects/games/pong/menu/InstructionsTextComponent.java
@@ -1,4 +1,6 @@
-package edu.ucsb.cs56.projects.games.pong;
+package edu.ucsb.cs56.projects.games.pong.menu;
+
+import edu.ucsb.cs56.projects.games.pong.menu.instructions.Instructions;
 
 import java.awt.*;
 import java.awt.event.MouseEvent;
