+++ src/edu/ucsb/cs56/projects/games/pong/menu/MainMenuComponent.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.projects.games.pong;
+package edu.ucsb.cs56.projects.games.pong.menu;
 
 import javax.swing.*;
 import java.awt.*;
@@ -6,7 +6,7 @@ import java.awt.*;
 /**
  * Created by angel on 2/20/16.
  */
-public abstract class MainMenuComponent extends JComponent implements UserRetrievable {
+public abstract class MainMenuComponent extends JComponent {
 
     private static Dimension dimension;
