+++ src/edu/ucsb/cs56/projects/games/pong/menu/MenuPaddingComponent.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.projects.games.pong;
+package edu.ucsb.cs56.projects.games.pong.menu;
 
 import java.awt.*;
 
@@ -6,21 +6,15 @@ import java.awt.*;
  * Created by angel on 2/20/16.
  */
 public class MenuPaddingComponent extends MainMenuComponent {
-    private final String title;
     private int height;
     private int width;
     private Color backgroundColor;
     public MenuPaddingComponent(Color backgroundColor,int height, int width) {
         super();
-        this.title = "MenuPadding";
         this.height = height;
         this.width = width;
         this.backgroundColor = backgroundColor;
         this.setPreferredSize(new Dimension(width,height));
-    }
-
-    public String getTitle() {
-        return title;
     }
 
     @Override
