+++ src/edu/ucsb/cs56/projects/games/pong/menu/MenuTextComponent.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.projects.games.pong;
+package edu.ucsb.cs56.projects.games.pong.menu;
 
 import java.awt.*;
 import java.awt.event.MouseEvent;
