--- src/edu/ucsb/cs56/projects/games/pong/scores.txt
@@ -1,1 +1,0 @@
-100/Throckmorton/75/Jenny/57/V /50/Richard/37//
